startup_eval_app/pyproject.toml
[project]
name = "startup-eval-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.1",
    "google-ai-generativelanguage>=0.6.15",
    "google-generativeai>=0.8.5",
    "gradio>=5.46.1",
    "matplotlib>=3.10.6",
    "mcp>=1.14.1",
    "pymupdf>=1.26.4",
]
